Adding 2 byte changes

+++ src/i2c.h
@@ -30,7 +30,7 @@
 *				FUNCTION PROTOTYPE
 *****************************************************************************************/
 bool i2c_write(uint8_t slave_address, uint8_t reg_addr,uint8_t* buffer, uint8_t num_bytes);
-bool i2c_read(uint8_t slave_address, uint8_t reg_addr,uint8_t* buffer,uint8_t num_bytes);
+bool i2c_read(uint8_t slave_address, uint8_t reg_addr,uint16_t* buffer,uint8_t num_bytes);
 
 
 #endif //I2C_H_
